feat: Add support for Automatic Tool Changer (ATC)
Introduces a `hasATC` flag in machine settings to differentiate between automatic and manual tool changes.

When ATC is enabled and a tool change (M6) is detected for a tool present in the library, the `automaticToolChange` script will be executed. Otherwise, the system will prompt for a manual tool change.

This change enhances the application's compatibility with machines equipped with an ATC system.